Added possibility to create relative links - with #...
It's very important in order to create Block of contents at the start of the article.

Probably we'll use for this ol/ul (list), in which relative links will be set.

The very first ol block in the article will have different styles and rendered differently to float aside with position sticky. The only way to do this is to use JS to move the block to another place, like teleport tag in Vue. Maybe there will be a way to catch that block and render it differently with PortableText, like Image or Link special types for Portabletext.